split logic from view controllers
